Set of Resource.localId from int to long

+++ src/resourcePackage/ImageResource.java
@@ -5,7 +5,7 @@ import java.net.URL;
 
 public class ImageResource extends Resource
 {
-	public ImageResource(int localId, URL url, File file, String name)
+	public ImageResource(long localId, URL url, File file, String name)
 	{
 		super(localId, url, file, name);
 	}
